@johannesvollmer 2023-12-05
Very interesting. I'm curious how it's possible to be faster than flume, is there something I can read about some nice performance tricks?  :)