@johannesvollmer 2023-12-05
Cool thanks :) I'll have a closer look at this tomorrow hopefully. Thanks so far :)