@mahdi-shojaee 2023-12-05
> Also, did you consider improving flume?

While I appreciate the suggestion, I'm currently prioritizing Loole development.